+++ R/sdplyr-Spatial.r
@@ -86,6 +86,7 @@ data_or_stop <- function(x, mess = "") {
   if (!.hasSlot(x, "data")) {
     stop("no data %s for a %s", mess, class(x))
   } 
+  if (inherits(x@data, "tbl_df")) return(x@data)
   as_tibble(x@data)
 }
 #' @export
@@ -129,7 +130,7 @@ mutate_.Spatial <-  function(.data, ..., .dots) {
 #' }
 summarise_.Spatial <- function(.data, ...) {
   dat <- data_or_stop(.data, " to summarize ")
-  dat <- summarise_(.data@data, ...)
+  dat <- summarise(dat, ...)
   
   # row.names(dat) <- "1"
   gbomb <- spbabel::sptable(.data)
@@ -160,17 +161,17 @@ summarise_.Spatial <- function(.data, ...) {
 summarise.Spatial <- function(.data, ...) {
   dat <- data_or_stop(.data, " to summarize ")
 
-  dat <- summarise(.data@data, ...)
+  dat <- summarise(dat, ...)
   
   # row.names(dat) <- "1"
   gbomb <- spbabel::sptable(.data)
   
   ## prepare the groups
   if (inherits(.data@data, "grouped_df")) {
-    groups <- attr(.data@data, "indices")  ## only robust for single-level group_by for now
-    grp_sizes <- attr(.data@data, "group_sizes")
+    groups <- dplyr::group_rows(.data@data)  ## attr(.data@data, "indices")  ## only robust for single-level group_by for now
+    grp_sizes <- dplyr::group_size(.data@data)
     regroup <- tibble(labs =  unlist(lapply(seq_along(grp_sizes), function(x) rep(x, grp_sizes[x]))), 
-                      inds = unlist(groups) + 1)
+                      inds = unlist(groups))
     
     gbomb <- gbomb  %>% 
       dplyr::inner_join(regroup, c("object_" = "inds"))  %>% 
@@ -189,7 +190,7 @@ summarise.Spatial <- function(.data, ...) {
 #' @name dplyr-Spatial
 #' @export
 group_by_.Spatial <- function(.data, ...) {
-  .data@data <- group_by_(data_or_stop(.data, " to group_by "), ...)
+  .data@data <- group_by(data_or_stop(.data, " to group_by "), ...)
   .data
 }